+++ pyproject.toml
@@ -111,6 +111,7 @@ known-first-party = [
   "shared_models",
   "text_to_code",
   "text_to_code_lambda",
+  "utils",
 ]
 
 [tool.ruff.lint.per-file-ignores]
